Fixed a lot of warnings

+++ AudioFile.h
@@ -306,7 +306,7 @@ bool AudioFile<T>::setAudioBuffer (AudioBuffer& newBuffer)
         return false;
     }
     
-    int numSamples = (int)newBuffer[0].size();
+    size_t numSamples = newBuffer[0].size();
     
     // set the number of channels
     samples.resize (newBuffer.size());
@@ -317,7 +317,7 @@ bool AudioFile<T>::setAudioBuffer (AudioBuffer& newBuffer)
         
         samples[k].resize (numSamples);
         
-        for (int i = 0; i < numSamples; i++)
+        for (size_t i = 0; i < numSamples; i++)
         {
             samples[k][i] = newBuffer[k][i];
         }
@@ -787,7 +787,7 @@ bool AudioFile<T>::saveToWaveFile (std::string filePath)
     }
     
     // check that the various sizes we put in the metadata are correct
-    if (fileSizeInBytes != (fileData.size() - 8) || dataChunkSize != (getNumSamplesPerChannel() * getNumChannels() * (bitDepth / 8)))
+    if (fileSizeInBytes != static_cast<int32_t> (fileData.size() - 8) || dataChunkSize != (getNumSamplesPerChannel() * getNumChannels() * (bitDepth / 8)))
     {
         reportError ("ERROR: couldn't save file to " + filePath);
         return false;
@@ -871,7 +871,7 @@ bool AudioFile<T>::saveToAiffFile (std::string filePath)
     }
     
     // check that the various sizes we put in the metadata are correct
-    if (fileSizeInBytes != (fileData.size() - 8) || soundDataChunkSize != getNumSamplesPerChannel() *  numBytesPerFrame + 8)
+    if (fileSizeInBytes != static_cast<int32_t> (fileData.size() - 8) || soundDataChunkSize != getNumSamplesPerChannel() *  numBytesPerFrame + 8)
     {
         reportError ("ERROR: couldn't save file to " + filePath);
         return false;
@@ -889,7 +889,7 @@ bool AudioFile<T>::writeDataToFile (std::vector<uint8_t>& fileData, std::string 
     
     if (outputFile.is_open())
     {
-        for (int i = 0; i < fileData.size(); i++)
+        for (size_t i = 0; i < fileData.size(); i++)
         {
             char value = (char) fileData[i];
             outputFile.write (&value, sizeof (char));
@@ -907,7 +907,7 @@ bool AudioFile<T>::writeDataToFile (std::vector<uint8_t>& fileData, std::string 
 template <class T>
 void AudioFile<T>::addStringToFileData (std::vector<uint8_t>& fileData, std::string s)
 {
-    for (int i = 0; i < s.length();i++)
+    for (size_t i = 0; i < s.length();i++)
         fileData.push_back ((uint8_t) s[i]);
 }
 
@@ -1024,7 +1024,7 @@ int AudioFile<T>::getIndexOfString (std::vector<uint8_t>& source, std::string st
         
         if (section == stringToSearchFor)
         {
-            index = i;
+            index = static_cast<int> (i);
             break;
         }
     }
