+++ tests/AudioFileTests/AiffLoadingTests.cpp
@@ -34,7 +34,7 @@ BOOST_AUTO_TEST_CASE (AiffLoadingTests_Stereo_8bit_44100)
     BOOST_CHECK_EQUAL (audioFile.getSampleRate(), aiff_stereo_8bit_44100::sampleRate);
     BOOST_CHECK_EQUAL (audioFile.getNumChannels(), aiff_stereo_8bit_44100::numChannels);
     
-    for (int i = 0; i < aiff_stereo_8bit_44100::testBuffer.size(); i++)
+    for (size_t i = 0; i < aiff_stereo_8bit_44100::testBuffer.size(); i++)
     {
         for (int k = 0; k < audioFile.getNumChannels(); k++)
         {
@@ -55,7 +55,7 @@ BOOST_AUTO_TEST_CASE (AiffLoadingTests_Stereo_16bit_44100)
     BOOST_CHECK_EQUAL (audioFile.getSampleRate(), aiff_stereo_16bit_44100::sampleRate);
     BOOST_CHECK_EQUAL (audioFile.getNumChannels(), aiff_stereo_16bit_44100::numChannels);
     
-    for (int i = 0; i < aiff_stereo_16bit_44100::testBuffer.size(); i++)
+    for (size_t i = 0; i < aiff_stereo_16bit_44100::testBuffer.size(); i++)
     {
         for (int k = 0; k < audioFile.getNumChannels(); k++)
         {
@@ -76,7 +76,7 @@ BOOST_AUTO_TEST_CASE (AiffLoadingTests_Stereo_24bit_44100)
     BOOST_CHECK_EQUAL (audioFile.getSampleRate(), aiff_stereo_24bit_44100::sampleRate);
     BOOST_CHECK_EQUAL (audioFile.getNumChannels(), aiff_stereo_24bit_44100::numChannels);
     
-    for (int i = 0; i < aiff_stereo_24bit_44100::testBuffer.size(); i++)
+    for (size_t i = 0; i < aiff_stereo_24bit_44100::testBuffer.size(); i++)
     {
         for (int k = 0; k < audioFile.getNumChannels(); k++)
         {
@@ -97,7 +97,7 @@ BOOST_AUTO_TEST_CASE (AiffLoadingTests_Stereo_8bit_48000)
     BOOST_CHECK_EQUAL (audioFile.getSampleRate(), aiff_stereo_8bit_48000::sampleRate);
     BOOST_CHECK_EQUAL (audioFile.getNumChannels(), aiff_stereo_8bit_48000::numChannels);
     
-    for (int i = 0; i < aiff_stereo_8bit_48000::testBuffer.size(); i++)
+    for (size_t i = 0; i < aiff_stereo_8bit_48000::testBuffer.size(); i++)
     {
         for (int k = 0; k < audioFile.getNumChannels(); k++)
         {
@@ -118,7 +118,7 @@ BOOST_AUTO_TEST_CASE (AiffLoadingTests_Stereo_16bit_48000)
     BOOST_CHECK_EQUAL (audioFile.getSampleRate(), aiff_stereo_16bit_48000::sampleRate);
     BOOST_CHECK_EQUAL (audioFile.getNumChannels(), aiff_stereo_16bit_48000::numChannels);
     
-    for (int i = 0; i < aiff_stereo_16bit_48000::testBuffer.size(); i++)
+    for (size_t i = 0; i < aiff_stereo_16bit_48000::testBuffer.size(); i++)
     {
         for (int k = 0; k < audioFile.getNumChannels(); k++)
         {
@@ -139,7 +139,7 @@ BOOST_AUTO_TEST_CASE (AiffLoadingTests_Stereo_24bit_48000)
     BOOST_CHECK_EQUAL (audioFile.getSampleRate(), aiff_stereo_24bit_48000::sampleRate);
     BOOST_CHECK_EQUAL (audioFile.getNumChannels(), aiff_stereo_24bit_48000::numChannels);
     
-    for (int i = 0; i < aiff_stereo_24bit_48000::testBuffer.size(); i++)
+    for (size_t i = 0; i < aiff_stereo_24bit_48000::testBuffer.size(); i++)
     {
         for (int k = 0; k < audioFile.getNumChannels(); k++)
         {
